+++ R/manager.R
@@ -1,7 +1,7 @@
 #' Manage Packages from System Repositories
 #'
 #' Talk to the system package manager to install/remove... packages from system
-#' repositories.
+#' repositories (see details for further options).
 #'
 #' @param pkgs character vector of names of packages.
 #' @return Functions \code{install_sys} and \code{remove_sys} return, invisibly,
@@ -16,6 +16,15 @@
 #' that, if you want to fall back to \code{sudo} in a non-interactive session,
 #' you need to set \code{options(bspm.sudo=TRUE)}.
 #'
+#' By default, if a package is not available in the system repositories, it is
+#' installed from R's configured repositories along with all its dependencies.
+#' This behavior can be changed via \code{options(bspm.always.install.deps=TRUE)},
+#' which tries to install from system repositories recursive dependencies of
+#' those packages that are not available. For example, if \pkg{A} depends on
+#' \pkg{B}, and \pkg{B} is available in the system repositories but \pkg{A} is
+#' not, then only \pkg{A} will be installed from CRAN with this option enabled,
+#' and both will be installed from CRAN with this option disabled (default).
+#'
 #' @seealso \code{\link{integration}}
 #'
 #' @examples
@@ -29,7 +38,15 @@
 #'
 #' @name manager
 #' @export
-install_sys <- function(pkgs) backend_call("install", pkgs)
+install_sys <- function(pkgs) {
+  not.avail <- backend_call("install", pkgs)
+  if (getOption("bspm.always.install.deps", FALSE)) {
+    deps <- tools::package_dependencies(not.avail, recursive=TRUE)
+    deps <- unique(unlist(deps, use.names=FALSE))
+    backend_call("install", deps)
+  }
+  not.avail
+}
 
 #' @name manager
 #' @export